+++ app/configuration/index.test.ts
@@ -295,7 +295,7 @@ describe('getSecurityConfiguration', () => {
     configuration.ddEnvVars.DD_SECURITY_COSIGN_IDENTITY = 'maintainer@example.com';
     configuration.ddEnvVars.DD_SECURITY_COSIGN_ISSUER = 'https://token.actions.githubusercontent.com';
     configuration.ddEnvVars.DD_SECURITY_SBOM_ENABLED = 'true';
-    configuration.ddEnvVars.DD_SECURITY_SBOM_FORMATS = 'cyclonedx,spdx-json,cyclonedx';
+    configuration.ddEnvVars.DD_SECURITY_SBOM_FORMATS = 'cyclonedx-json,spdx-json,cyclonedx-json';
 
     const result = configuration.getSecurityConfiguration();
     expect(result).toEqual({
@@ -319,7 +319,7 @@ describe('getSecurityConfiguration', () => {
       },
       sbom: {
         enabled: true,
-        formats: ['cyclonedx', 'spdx-json'],
+        formats: ['cyclonedx-json', 'spdx-json'],
       },
     });